addressed feedback from @somanyhs
